fix: Pattern matching
If the left expression is a list with a symbol binding and the right
value in the same position is a marker pair with a differe marker, it
should not match.

e.g.:

(set [x] [1])

should succeed and bind x to 1, but

(set [first-name] [:last-name "Doe"])

should fail.

Before it was just quietly succeeding without binding any variables.